changing sample age to double

+++ src/edu/mcw/rgd/dao/spring/GeneExpressionQuery.java
@@ -49,8 +49,8 @@ public class GeneExpressionQuery extends MappingSqlQuery {
 
         Sample s = new Sample();
         s.setId(rs.getInt("sample_id"));
-        s.setAgeDaysFromHighBound(rs.getInt("age_days_from_dob_high_bound"));
-        s.setAgeDaysFromLowBound(rs.getInt("age_days_from_dob_low_bound"));
+        s.setAgeDaysFromHighBound(rs.getDouble("age_days_from_dob_high_bound"));
+        s.setAgeDaysFromLowBound(rs.getDouble("age_days_from_dob_low_bound"));
         s.setNumberOfAnimals(rs.getInt("number_of_animals"));
         s.setNotes(rs.getString("sample_notes"));
         s.setSex(rs.getString("sex"));
